refactor: share the subprocess spawn/timeout/temp-cwd core across runners
The plan's refactor step for Slice 8. The spawn/drain/timeout/process-group-
kill/temp-cwd dance was identical in r.rs and python.rs; extract it into a new
private `runner::subprocess` module. Each runner is now a small `Interpreter`
descriptor (program + the args preceding the code) over the shared `run`
(§4.2 — factored, not a per-language copy and not a utils dump). Adding a
language is a descriptor, not a re-implementation of the mechanics (§3.2/§3.4).

RunnerError's context widens from `&'static str` to `String` (via
`impl Into<String>`) so the shared core can name the failing interpreter
(`format!("spawn {program}")`) instead of regressing R's "spawn Rscript"
diagnostics to a generic label. All existing call sites are `&str` literals
that coerce unchanged; swept the whole crate (only the helper and the
RunnerError unit test construct it).

No behavior change: full suite 42/42 green (R AC1-3 + Python AC1-2, 0 skipped),
clippy -D warnings clean, rustdoc -D warnings clean.
