+++ crates/core/src/runner/r.rs
@@ -1,19 +1,22 @@
-//! R execution mechanics: spawn `Rscript`, capture its streams separately,
-//! bound it with a timeout, and reap a runaway process tree.
+//! R execution mechanics: the R-specific half of a language run — which program
+//! and which flags — over the shared subprocess core in [`super::subprocess`].
 //!
-//! This module owns *only* the R subprocess details (ADR-0005, §4.1); callers
-//! depend on the [`Runner`](super::Runner) trait, not on [`RRunner`].
+//! This module owns *only* the R specifics (ADR-0005, §4.1); callers depend on
+//! the [`Runner`](super::Runner) trait, not on [`RRunner`], and the
+//! spawn/timeout/temp-cwd dance is shared with every other language (§4.2).
 
-use std::process::Stdio;
 use std::time::Duration;
 
-use command_group::AsyncCommandGroup;
-use tokio::io::AsyncReadExt;
-use tokio::process::Command;
-use tokio::task::JoinHandle;
-use tokio::time::timeout;
-
+use super::subprocess::{self, Interpreter};
 use super::{ExecutionResult, Runner, RunnerError, Timeout};
+
+/// The R invocation: `Rscript --vanilla -e <code>`. `--vanilla` skips user/site
+/// `.Rprofile`/`.Renviron`, so capture is deterministic across machines and no
+/// profile output bleeds into stdout.
+const R_INTERPRETER: Interpreter = Interpreter {
+    program: "Rscript",
+    code_args: &["--vanilla", "-e"],
+};
 
 /// A [`Runner`] backed by a real `Rscript` subprocess.
 #[derive(Debug, Clone)]
@@ -41,81 +44,6 @@ impl Runner for RRunner {
         code: &str,
         _checks: &[String],
     ) -> Result<ExecutionResult, RunnerError> {
-        // Per-execute temp dir as CWD: file-writing code is isolated to it and
-        // the directory is removed when `run_dir` drops at the end of this fn.
-        let run_dir =
-            tempfile::TempDir::new().map_err(|e| RunnerError::new("create run directory", e))?;
-
-        // `--vanilla` skips user/site `.Rprofile`/`.Renviron`, so capture is
-        // deterministic across machines and no profile output bleeds into stdout.
-        // `group_spawn` puts the child in its own process group so the whole tree
-        // can be killed on timeout, not just the leader.
-        let mut child = Command::new("Rscript")
-            .arg("--vanilla")
-            .arg("-e")
-            .arg(code)
-            .current_dir(run_dir.path())
-            .stdin(Stdio::null())
-            .stdout(Stdio::piped())
-            .stderr(Stdio::piped())
-            .group_spawn()
-            .map_err(|e| RunnerError::new("spawn Rscript", e))?;
-
-        // Drain both pipes concurrently so a chatty program cannot deadlock the
-        // wait by filling a pipe buffer; the reads make progress while we await
-        // the process below.
-        let mut stdout = child.inner().stdout.take().expect("stdout piped above");
-        let mut stderr = child.inner().stderr.take().expect("stderr piped above");
-        let out_task: JoinHandle<std::io::Result<Vec<u8>>> = tokio::spawn(async move {
-            let mut buf = Vec::new();
-            stdout.read_to_end(&mut buf).await.map(|_| buf)
-        });
-        let err_task: JoinHandle<std::io::Result<Vec<u8>>> = tokio::spawn(async move {
-            let mut buf = Vec::new();
-            stderr.read_to_end(&mut buf).await.map(|_| buf)
-        });
-
-        let (exit, timed_out) = match timeout(self.timeout.0, child.wait()).await {
-            Ok(status) => {
-                let status = status.map_err(|e| RunnerError::new("wait for Rscript", e))?;
-                (status.code(), false)
-            }
-            Err(_elapsed) => {
-                // SIGKILL the whole process group so an R child that spawned its
-                // own children is reaped, not leaked; then reap the leader.
-                child
-                    .kill()
-                    .await
-                    .map_err(|e| RunnerError::new("kill Rscript group", e))?;
-                let status = child
-                    .wait()
-                    .await
-                    .map_err(|e| RunnerError::new("reap Rscript group", e))?;
-                (status.code(), true)
-            }
-        };
-
-        // Every group process is gone, so both pipes have closed and the readers
-        // have finished; collect their bytes.
-        let stdout = join_capture(out_task, "collect stdout").await?;
-        let stderr = join_capture(err_task, "collect stderr").await?;
-
-        Ok(ExecutionResult::from_capture(
-            &stdout, &stderr, exit, timed_out,
-        ))
-        // `run_dir` drops here → the temp directory is removed.
-    }
-}
-
-/// Await a stream-reader task, flattening its join error and IO error into a
-/// [`RunnerError`].
-async fn join_capture(
-    task: JoinHandle<std::io::Result<Vec<u8>>>,
-    context: &'static str,
-) -> Result<Vec<u8>, RunnerError> {
-    match task.await {
-        Ok(Ok(bytes)) => Ok(bytes),
-        Ok(Err(io_err)) => Err(RunnerError::new(context, io_err)),
-        Err(join_err) => Err(RunnerError::new(context, std::io::Error::other(join_err))),
+        subprocess::run(&R_INTERPRETER, code, self.timeout).await
     }
 }
